大まかなベース作成完了
